move performance month graph generation to background job
